perf(tool-proxy): 5.5s -> 0.7s pod startup — stop blocking boot on an unreachable beacon
Measured, before and after, on a real Firecracker pod on the M5 (Lima vz -> KVM
-> Firecracker 1.16.1):

    BEFORE   boot_report=5044ms   total=5497ms   harness: health check timed out
    AFTER    boot_report= 266ms   total= 691ms   harness: CREATE 200

## The 6.1 seconds was one function

#2175 localised ~6.1 s of a 7.7 s pod create to "guest tool-proxy startup" and
could go no further: the guest proxy logs nothing. Instrumenting it (previous
commit) put 92% of the startup in a single call — `emit_boot_report`, at 5044 ms
of 5497 ms.

`emit_boot_report` writes the first audit entry. `AuditLog::log` fills
`drand_round` by fetching `https://api.drand.sh/public/latest` with a 5 s
timeout (`nucleus_client::drand`, `.timeout(Duration::from_secs(5))`).
`drand_enabled` defaults to `true`.

A pod runs under **default-deny egress**. The fetch cannot succeed — not
"usually fails", cannot. So every pod paid a five-second timeout for a network
round trip its own network policy forbids, before serving a single request. The
cost was invisible because it is one `.await` in a 670-line startup with no
timing on it.

## The fix

`emit_boot_report_bounded`: a 250 ms deadline. Long enough for a beacon that is
genuinely reachable (an operator who allows the drand host in `dns_allow`), far
short of the doomed case. On expiry the entry is still written — the hash chain
must not gain a hole — just without an anchor it was never going to get.

**This does not weaken approvals.** Anchoring exists to stop pre-computation of
approval SIGNATURES; `/v1/approve` validates its own round, and
`DrandFailMode::Strict` still refuses an approval when the beacon is absent.
What changes is that the refusal is no longer paid at boot by every pod.

## Two environment bugs found on the way, both already fixed in main

Reaching a healthy boot at all needed current binaries, and the failures were
informative:

* the July node binary never created `vsock.sock_15012`, so the guest's
  workload-API connect was reset — the pod dir had only `vsock.sock`. Current
  main starts `WorkloadApiVsockBridge` before the health wait (main.rs:2743 vs
  2828, with a comment saying it used to be the other way round). Already fixed.
* the July guest-init demanded `nucleus.approval_secret`, which current main
  correctly no longer injects (replaced by `approval_pubkeys` in the 2026-08-08
  cmdline cleanup that `snapshot.rs` documents). PID 1 exited; kernel panic.

Neither is a defect in current code. Both are why the measurement needed
current node + current guest-init + instrumented proxy in one rootfs.

## Ratchet

`emit_boot_report` and its bounded wrapper moved to `boot_report.rs`.
main.rs 5010 -> 4944, ceiling 4975.

  nucleus-tool-proxy 358 passed · clippy --all-features -D warnings clean · fmt clean
